Switching to spring boot

+++ src/main/java/com/example/model/movie/Film.java
@@ -4,12 +4,13 @@ import com.example.jackson.LocalDateDeserializer;
 import com.example.jackson.LocalDateSerializer;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
+import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import javax.persistence.*;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.Set;
 
@@ -38,7 +39,7 @@ public class Film {
     private LocalDate releaseDate;
 
     @Column(name = "rating", precision = 2, scale = 1)
-    private Double rating;
+    private BigDecimal rating;
 
     @Column(name = "description", columnDefinition = "TEXT")
     private String description;
